+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Fix formatting for trailing comments at the end of block statements ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/12)).
 - Fix issue in which comments right before semi colon are not printed ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/11)).
 - Assert that all comments have been printed out in the formatted code.
+- Fix long expressions not breaking into multi lines ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/16)).
 
 ## 1.0.0-alpha.5