+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 - Fix Annotated Declaration being indented too far ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/18)).
+- Getter and setter will no longer break if they can fit on the same line.
 
 ## 1.0.0-alpha.6
 - Fix formatting for trailing comments at the end of block statements ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/12)).