+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 - Fix Annotated Declaration being indented too far ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/18)).
 - Getter and setter will no longer break if they can fit on the same line.
+- Add documentation for configuration options.
 
 ## 1.0.0-alpha.6
 - Fix formatting for trailing comments at the end of block statements ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/12)).